requirements.txt
fastapi
pydantic
pydantic-settings
pydantic_core
pymongo
python-dotenv
uvicorn
starlette
fastapi-versioning
beanie
pytest
httpx

asgi_lifespan
mongomock
mongomock-motor
pytest-asyncio
black
passlib[bcrypt]
fastapi-mail
Jinja2
mailtrap